feat(submitentityclaim): refactor to make controls generic
