+++ src/common/components/JsonForm/CustomWidgets/ImageCheckboxes/ImageCheckboxes.tsx
