+++ src/common/components/JsonForm/SingleControlForm/SingleControlForm.styles.ts
@@ -1,0 +1,11 @@
+import styled from 'styled-components'
+
+export const FormContainer = styled.div`
+  form {
+    width: 100%;
+
+    .buttons {
+      float: right;
+    }
+  }
+`
