+++ src/common/components/JsonForm/types.ts
@@ -1,0 +1,15 @@
+export interface FormControl {
+  id: string
+  title: string
+  description: string
+  required: boolean
+  type: string
+  label: string
+  control: string
+  placeholder?: string
+  itemIds?: string[]
+  itemLabels?: string[]
+  itemImages?: string[]
+  minItems?: number
+  maxItems?: number
+}
