+++ src/modules/SubmitEntityClaim/SubmitEntityClaim.selectors.ts
@@ -1,6 +1,7 @@
 import { createSelector } from 'reselect'
 import { RootState } from 'src/common/redux/types'
-import { SubmitEntityClaimState, Question } from './types'
+import { SubmitEntityClaimState } from './types'
+import { FormControl } from '../../common/components/JsonForm/types'
 
 export const selectSubmitEntityClaim = (
   state: RootState,
@@ -23,14 +24,14 @@ export const selectCurrentQuestionNo = createSelector(
 export const selectCurrentQuestion = createSelector(
   selectQuestions,
   selectCurrentQuestionNo,
-  (questions: Question[], currentQuestionNo: number) => {
+  (questions: FormControl[], currentQuestionNo: number) => {
     return questions[currentQuestionNo - 1]
   },
 )
 
 export const selectQuestionCount = createSelector(
   selectQuestions,
-  (questions: Question[]) => {
+  (questions: FormControl[]) => {
     return questions.length
   },
 )
