--- src/modules/SubmitEntityClaim/components/Question/Question.styles.ts
@@ -1,11 +1,0 @@
-import styled from 'styled-components'
-
-export const QuestionContainer = styled.div`
-  form {
-    width: 100%;
-
-    .buttons {
-      float: right;
-    }
-  }
-`
