+++ src/modules/SubmitEntityClaim/temp_questions.json
@@ -1,6 +1,6 @@
 [
   {
-    "questionId": "a",
+    "id": "a",
     "title": "Short Text Question",
     "description": "This will be a short description or explainer text explaining the question. ",
     "label": "Short Answer",
@@ -10,7 +10,7 @@
     "placeholder": "Start Typing here"
   },
   {
-    "questionId": "b",
+    "id": "b",
     "title": "Long Text Question",
     "description": "This will be a short description or explainer text explaining the question. ",
     "label": "Long Answer",
@@ -20,7 +20,7 @@
     "placeholder": "Start Typing here"
   },
   {
-    "questionId": "c",
+    "id": "c",
     "title": "Select some of the text items below",
     "description": "This will be a short description or explainer text explaining the question. ",
     "label": "Select 1 to 2 options",
@@ -49,7 +49,7 @@
     ]
   },
   {
-    "questionId": "d",
+    "id": "d",
     "title": "Select some of the image items below",
     "description": "This will be a short description or explainer text explaining the question. ",
     "label": "Select 3 options",
@@ -57,7 +57,7 @@
     "type": "array",
     "minItems": 3,
     "maxItems": 3,
-    "control": "custom_imagecheckboxes",
+    "control": "imagecheckboxes",
     "itemIds": ["Option1", "Option2", "Option3", "Option4"],
     "itemLabels": [
       "Picture Option 1",
