+++ src/modules/SubmitEntityClaim/types.ts
@@ -1,21 +1,7 @@
-export interface Question {
-  questionId: string
-  title: string
-  description: string
-  required: boolean
-  type: string
-  label: string
-  control: string
-  placeholder?: string
-  itemIds?: string[]
-  itemLabels?: string[]
-  itemImages?: string[]
-  minItems?: number
-  maxItems?: number
-}
+import { FormControl } from '../../common/components/JsonForm/types'
 
 export interface SubmitEntityClaimState {
-  questions: Question[]
+  questions: FormControl[]
   currentQuestionNo: number
   answers: any[]
   answersComplete: boolean
